refactor(frontend): history store -> history helper
